pyproject.toml
[project]
name = "fbi-dwh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.0",
    "geopandas>=1.0.1",
    "jupyter>=1.1.1",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "requests>=2.32.3",
    "ruff>=0.9.6",
    "streamlit>=1.42.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
exclude = [
    ".env",
    ".git",
    "*.ipynb",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".venv",
]

line-length = 100
indent-width = 4
